Change default MemoryOutStream to be a std::ostringstream.
This changes restores the default behavior to that of UT++ 1.4.

It is accomplished by reintroducing the old UNITTEST_USE_CUSTOM_STREAMS
define, commented out, and surrounding the new define in an ifndef
block. It was done this way to preserve any unconventional users that
were specifying either flag in their build system, instead of in this
file.